error cicd fix;

+++ controllers/courses/requests/updateCourse.go
@@ -6,9 +6,9 @@ type UpdateCourse struct {
 	Title       string `json:"title"`
 	Description string `json:"description"`
 	FileImage   string `json:"file_image"`
-	Status int `json:"status"`
+	Status int8 `json:"status"`
 }
 
 func (c UpdateCourse) ToDomain() *courses.Domain {
-	return &courses.Domain{Title: c.Title, Description: c.Description, ImgUrl: c.FileImage, Status: int8(c.Status)}
+	return &courses.Domain{Title: c.Title, Description: c.Description, ImgUrl: c.FileImage, Status: c.Status}
 }
